Add initial draft of functions
This commit adds the basic functions and documentation for the
raincheck package: a function `raincheck` that creates a function
that will execute in the environment where it is called, and a
fuction `scold` that will issue warnings and errors that spoof the
calling environment, so that it can give more informative error
messages to end-users.

The commit also makes the README informative and starts package
documentation.